recycle: Refactor drivers import in connection.component.ts
The import statements for the drivers module in the connection.component.ts file have been refactored to use the updated file structure.

:art: Format code for base editor initialization

The code for initializing the base editor in the trigger.component.html file has been formatted to improve readability and maintainability.

:pencil: Update import path for Cockroach driver

The import path for the Cockroach driver in the cockroach.ts file has been updated to reflect the new file structure.

:pencil: Update import path for ElasticSearch driver

The import path for the ElasticSearch driver in the elasticsearch.ts file has been updated to reflect the new file structure.

:pencil: Update import path for MongoDB driver

The import path for the MongoDB driver in the mongodb.ts file has been updated to reflect the new file structure.

:pencil: Update import path for MySQL driver

The import path for the MySQL driver in the mysql.ts file has been updated to reflect the new file structure.

:pencil: Update import path for Oracle driver

The import path for the Oracle driver in the oracle.ts file has been updated to reflect the new file structure.

:pencil: Update import path for PostgreSQL driver

The import path for the PostgreSQL driver in the postgresql.ts file has been updated to reflect the new file structure.

:pencil: Update import path for SQL driver

The import path for the SQL driver in the sql.ts file has been updated to reflect the new file structure.

:pencil: Update import path for SQLite driver

The import path for the SQLite driver in the sqlite.ts file has been updated to reflect the new file structure.

:pencil: Update import path for SqlServer driver

The import path for the SqlServer driver in the sqlserver.ts file has been updated to reflect the new file structure.

:recycle: Refactor import path for helper.ts

The import path for the SQL class in the helper.ts file has been refactored to reflect the new file structure.

:recycle: Refactor drivers import in request.service.ts

The import statements for the drivers module in the request.service.ts file have been refactored to use the updated file structure.