sparkles: Added Fct class to handle function objects
The `Fct` class is added to the project in order to handle function objects. This class includes properties such as `code`, `saved`, `name`, `hide`, and `changed`. It provides a structured representation of functions and their associated attributes. This change is made to improve the organization and management of functions within the project.