art: Update container component styles
Changed the color of selected database name to teal (#1de9b6) for improved visibility. This change was made to enhance the user experience and make the selected database more distinguishable.

:art: Update container component styles

Changed the color of tab links in the mat-drawer to teal (#1de9b6) for improved visibility. This change was made to enhance the user experience and make the tab links more distinguishable.

:art: Update advanced component styles

Changed the color of material symbols from #ff9800 to teal (#1de9b6) for improved visibility and consistency. This change was made to enhance the user experience and make the material symbols more distinguishable.

:art: Update insert component styles

Changed the style of the monaco editor to flex: 1 for better resizing behaviour. This change was made to enhance the user experience and make the monaco editor more responsive.

:bug: Fix duplicate database button color

Changed the color of the duplicate database button to teal (#1de9b6) for consistency with other buttons. This change was made to enhance the user experience and make the buttons more cohesive.

:art: Update relations component styles

Changed the border colors of expanded rows in the table to teal (#1de9b6) for improved visibility and consistency. This change was made to enhance the user experience and make the expanded rows more distinguishable.

:art: Update column component styles

Changed the color of column type options to teal (#1de9b6) for improved visibility and consistency. This change was made to enhance the user experience and make the column type options more distinguishable.

:art: Update styles

Changed the color of anchor links to teal (#1de9b6) for improved visibility. This change was made to enhance the user experience and make the anchor links more distinguishable.