art: Rearrange buttons in container.component.html
The buttons in the container.component.html file have been rearranged to improve the user interface and make the navigation more intuitive. The "drawer.toggle()" function has been moved from the "click" event of the <nav> element to the "click" event of each <a> element within the <nav> element.

:bug: Fix padding of toolbar in query.component.html

The padding of the <mat-toolbar> element in the query.component.html file has been fixed. The inline style "padding: 10px 20px;" has been replaced with a class "toolbar" defined in the query.component.scss file.

:recycle: Remove unnecessary styling in query.component.scss

Unnecessary styling for the "#actions" ID selector in the query.component.scss file has been removed.

:bug: Fix model markers in query.component.ts

The model markers in the query.component.ts file have been fixed. Instead of calling "monaco.editor.setModelMarkers" on a single editor model, it is now called on each editor model in the "this.editors" array.

:sparkles: Add buttons to run and copy AI code in ai.component.html

Buttons to run and copy the AI code have been added to the ai.component.html file. The buttons have the "color" attribute set to "accent" for visual consistency.

:sparkles: Add buttons in history.component.html

Buttons for playing, starring, copying, and removing history items have been added to the history.component.html file. The buttons have the "color" attribute set to "accent" for visual consistency.