art: Update diagram hoverClass colorscss
Change the background and stroke color of the `.hoverClass` in the diagram component to RGBA 29, 233, 182 with an opacity of 0.8.

The changes were done to improve the visual appearance of the hover effect in the diagram component.

:bug: Update stroke color for path in diagram class

Change the stroke color of the path element in the diagram component to RGBA 29, 233, 182.

The changes were done to fix a bug where the stroke color was incorrect for the path element in the diagram component.

:package: Update README.md

Remove unnecessary information about cloud services and add additional information about unit testing in front-end and function+type in column/autocomplete.

The changes were done to provide more relevant and accurate information in the README.md file.