art: Refactor mongodb driver and structure GUI components
Refactor the `mongodb` driver by adding a `getColumns()` method to simplify the code. Update the `getBaseDelete()`, `getBaseInsert()`, and `getBaseSelect()` methods in the driver to use the `getColumns()` method.

Refactor the `structure.component.html` file by removing the link to the NoSQL disclaimer and adding a button to add a column.

Refactor the `config-dialog.component.ts` file by only reloading the server if a server is selected before updating the configuration.

Refactor the `top-right.component.html` file by removing the link to the NoSQL disclaimer and adding a button for the Disclaimer.

Add a new configuration option `noSqlSample` to the `configuration.ts` file, which represents the sample size to infer NoSQL structure from. It has values '50', '100', '250', and '500'.

Reason for the changes:
- Refactoring the `mongodb` driver improves code readability and maintainability.
- Removing the link to the NoSQL disclaimer and adding a button for adding a column in the structure GUI components provides a better user experience and simplifies UI navigation.
- Reloading the server only when a server is selected after updating the configuration prevents unnecessary server reloading.
- Adding the `noSqlSample` configuration option allows users to configure the sample size used to infer NoSQL structure.