recycle: Improve MongoDB import command
The `mongoimport` command in the `MongoDB` wrapper has been updated to include the `--collection` flag and specify the collection name based on the original file name. This improvement allows for more efficient and precise data import into MongoDB databases.

The changes were done to enhance the functionality of the `MongoDB` wrapper and improve the process of importing data into MongoDB databases.