sparkles: Add changes for README.md and back/src/wrapper/mongodb.js
- Updated the Bin section in README.md to include information about dependencies such as Brew, Chocolate, Apt, and Yum. Also added a link to a pull request in the Electron repository and mentioned its use in CI and GitLab wiki.
- Made changes to the getRelations function in mongodb.js. Removed a comment, added new variables for sample size and minimum correlations, and made modifications to the code logic to improve the retrieval of relations between tables in databases.

WHY: The changes in the README.md file were made to provide more comprehensive information about the dependencies used in the project and to improve documentation. The changes in mongodb.js were made to enhance the functionality of retrieving relations between tables in databases by optimizing the code logic and variables used. This improves the reliability and accuracy of the relations retrieved, leading to better overall performance.