🚀 Added button to export all results
Added a button in the query component's HTML template to allow users to export all the results. The button is disabled if there is no data to export.

🔄 Updated autoFormat configuration

In the query component's TypeScript file, the `autoFormat` configuration value was updated to use the value from the configuration service. This ensures that the `autoFormat` setting is retrieved correctly.

WHY:
The changes were made to enhance the functionality of the query component. The new button provides users with the ability to export all the results, improving the usability of the component. Additionally, updating the `autoFormat` configuration ensures that the correct value is used for the `autoFormat` setting.