recycle: Remove unnecessary button and fix issue creation link
The unnecessary button for the user menu was removed from the top right component HTML file. In addition, the issue creation link in the logs-dialog HTML file was fixed to only include the first 1000 characters of the log message. This was done to ensure that the link does not exceed the allowed length for issue descriptions.

These changes were made to improve the user experience by removing redundant elements and fixing a bug in the issue creation feature.