art: Remove unnecessary attribute in container.component.html
The "matTooltip" attribute was removed from the button element in the container component. This was done to improve the code and remove unnecessary tooltips on the button.

:recycle: Refactor explore.component.html
Refactored the explore.component.html file to improve code readability and maintainability. Removed the unnecessary "hidePageSize" attribute from the mat-paginator element and added the [pageSizeOptions] attribute with an array of available page sizes.

:recycle: Refactor explore.component.ts
Refactored the explore.component.ts file to improve code readability and maintainability. Updated the pageSize assignment from the configuration file to a hardcoded value of 50.

:art: Remove unnecessary attribute in insert.component.ts
The "found" variable was removed from the sampleData method in the insert.component.ts file. This variable was not being used and was unnecessary.

:art: Remove unnecessary attribute in query.component.html
The "pageSize" attribute was removed from the mat-paginator element in the query.component.html file. This attribute was not needed as the [pageSize] attribute already sets the default page size.

:recycle: Refactor query.component.ts
Refactored the query.component.ts file to improve code readability and maintainability. Updated the pageSize assignment from the configuration file to a hardcoded value of 50.

:recycle: Refactor logs-dialog.html
Refactored the logs-dialog.html file to improve code readability and maintainability. Removed the unnecessary mat-progress-bar element and added a ViewChild reference to the scroll container.

:recycle: Refactor top-right.component.ts
Refactored the top-right.component.ts file to improve code readability and maintainability. Added a ViewChild decorator to get a reference to the scroll container and added a new ngOnInit method to scroll to the bottom of the container after loading the logs.

:recycle: Refactor column.component.html
Refactored the column.component.html file to improve code readability and maintainability. Removed the unnecessary mat-card-header elements and added the mat-form-field elements directly inside the mat-card-content elements for both the old and formColumn controls.

:recycle: Refactor styles.scss
Refactored the styles.scss file to improve code readability and maintainability. Made various styling adjustments to the mat-paginator, mat-paginator-container, and mat-paginator-range-label elements to align the page size dropdown on the right side and update the font size and alignment of the range label.