sparkles: Create a new table
Updating the `createTable` method in `mongodb.js` to create a new table with the specified name in the given database. The method inserts a new row into the table with the default values for each column.

:recycle: Refactor table creation dialog

In the `create-table-dialog.html` file, updated the title of the dialog to display the name of the selected database. This provides better context to the user when creating a new table.

:art: Improve column component layout

In the `column.component.html` file, refactored the layout of the column component. Moved the default value field and extras field to separate divs for better organization and styling.

:bug: Fix default value and extras fields

In the `column.ts` file, fixed the default value and extras fields to handle correct form control values. Updated the FormControl creation to correctly handle nullable and default value properties.

:recycle: Refactor shared module imports

In the `shared.module.ts` file, refactored the imports for the shared module to be more readable and organized. Added the missing `MatDividerModule` import to allow for the use of dividers in the UI.