✨ README.md: refactor: Harmonize column types with array > 2
Refactor the column types in the README.md file to harmonize them when they
contain arrays with more than 2 elements. This is necessary to maintain
consistency and avoid issues when using the aggregation lookup operator with
an array. The change was made to ensure reliable functionality and improve
data integrity.

🔨 back/src/wrapper/postgre.js: refactor: Update object and array columns

Refactor the `insert` method in the `postgre.js` file to update the handling of
object and array columns. Instead of wrapping the values in quotes, the method
now converts the objects and arrays to strings using `JSON.stringify()`. This
change ensures that the values are stored correctly in the database.

This change was made to improve the compatibility and consistency of the
insertion process in PostgreSQL.

🔨 front/src/app/core/insert/insert.component.html: update: Add stringify attribute

Add the `stringify` attribute to the app-cell component in the
insert.component.html file. This attribute is used to determine whether to
stringify the value before displaying it. The change ensures that object and
array values are properly displayed in the insert component.

This change was made to improve the user experience and ensure the correct
rendering of complex values in the insert component.

🔨 front/src/app/core/insert/insert.component.ts: refactor: Remove redundant code

Remove the redundant code that was stringifying the random generated value in
the insert.component.ts file. The value is already stringified in the
postgre.js file, so the additional `JSON.stringify()` call is unnecessary.

This change was made to simplify the code and improve performance.

🔨 front/src/app/tables/tables.component.html: update: Change placeholder text

Change the placeholder text in the tables.component.html file to include the
"Type" keyword in addition to "Table" and "Column". This change provides
clarification and improves the usability of the search feature.

This change was made to enhance the user experience and make the search more
intuitive.

🔨 front/src/app/tables/tables.component.ts: refactor: Apply search to column types

Update the search functionality in the tables.component.ts file to include the
column types in the search. Previously, only the table names and column names
were considered in the search. With this change, the search now includes the
column types as well, providing more comprehensive results.

This change was made to improve the search functionality and make it more
useful for users.