🐛 Fix: add gitlab-ci configuration for git package version
Added the necessary configuration in `.gitlab-ci.yml` file to install `git` package before running `eslint` job. This change was made to ensure that the `sed` command in `before_script` could function properly, as it requires `git` to retrieve the package version.

📝 WHY:
The `before_script` in the `.gitlab-ci.yml` file needed to install the `git` package so that the `sed` command could replace the `PACKAGE_VERSION` placeholder in the `package.json` file. This change was necessary for the proper functioning of the CI/CD pipeline.