✨ ️GitMoji: :recycle: Refactor
Update package version in .gitlab-ci.yml

The package version in the .gitlab-ci.yml file is being updated to use the branch name followed by the count of commits in the branch. This change is necessary to ensure that the correct version is used during deployment. The update is made using the sed command and the git rev-list --count HEAD command.